fixed the routing issues for invoice
