+++ careconnect2025/frontend/lib/config/navigation/caregiver-more-features-bottom-drawer.dart
@@ -1,4 +1,4 @@
-import 'package:care_connect_app/features/invoices/screens/invoice_dashboard_page.dart';
+import 'package:care_connect_app/features/invoices/screens/dashboard/invoice_dashboard_page.dart';
 import 'package:care_connect_app/pages/notetaker_configuration_page.dart';
 import 'package:care_connect_app/pages/settings_page.dart';
 import 'package:care_connect_app/shared/widgets/more_features_bottom_drawer.dart';
