+++ careconnect2025/frontend/lib/config/router/app_router.dart
@@ -4,7 +4,6 @@ import 'package:care_connect_app/features/integrations/presentation/pages/medica
 import 'package:care_connect_app/features/integrations/presentation/pages/smart_devices.dart';
 import 'package:care_connect_app/features/integrations/presentation/pages/wearables_screen.dart';
 import 'package:care_connect_app/features/invoices/screens/invoice_tabbed_page.dart';
-import 'package:care_connect_app/features/invoices/screens/upload_invoice.dart';
 import 'package:care_connect_app/features/profile/presentation/pages/profile_settings_page.dart';
 import 'package:care_connect_app/features/tasks/presentation/assign_task_screen.dart';
 import 'package:care_connect_app/features/tasks/presentation/calendar_assisiant.dart';
@@ -52,7 +51,7 @@ import '../../features/evv/presentation/pages/evv_visit_history.dart';
 import '../../features/evv/presentation/pages/evv_corrections.dart';
 import '../../features/evv/presentation/pages/evv_offline_sync.dart';
 import '../../providers/user_provider.dart';
-import 'package:care_connect_app/features/invoices/screens/invoice_dashboard_page.dart';
+import 'package:care_connect_app/features/invoices/screens/dashboard/invoice_dashboard_page.dart';
 import 'package:care_connect_app/features/invoices/screens/invoice_detail_page.dart';
 import 'package:care_connect_app/features/invoices/screens/invoice_list_page.dart';
 import 'package:care_connect_app/features/invoices/models/invoice_models.dart';
@@ -748,13 +747,15 @@ final GoRouter appRouter = GoRouter(
         GoRoute(
           path: 'list',
           name: 'invoiceList',
-         builder: (context, state) => const InvoiceTabbedPage(initialTabIndex: 2),
+          builder: (context, state) => const InvoiceTabbedPage(initialTabIndex: 2),
           routes: [
             GoRoute(
               path: ':filter',
               name: 'invoiceListFiltered',
-              builder: (context, state) =>
-                  InvoiceListPage(quickFilter: state.pathParameters['filter']),
+              builder: (context, state) => InvoiceTabbedPage(
+                initialTabIndex: 2,
+                quickFilter: state.pathParameters['filter'],
+              ),
             ),
           ],
         ),
