+++ careconnect2025/frontend/lib/features/invoices/screens/invoice_tabbed_page.dart
@@ -1,12 +1,17 @@
 import 'package:flutter/material.dart';
-import 'package:care_connect_app/features/invoices/screens/invoice_dashboard_page.dart';
+import 'package:care_connect_app/features/invoices/screens/dashboard/invoice_dashboard_page.dart';
 import 'package:care_connect_app/features/invoices/screens/upload_invoice.dart';
 import 'package:care_connect_app/features/invoices/screens/invoice_list_page.dart';
 
 class InvoiceTabbedPage extends StatefulWidget {
-  const InvoiceTabbedPage({super.key, this.initialTabIndex = 0});
+  const InvoiceTabbedPage({
+    super.key,
+    this.initialTabIndex = 0,
+    this.quickFilter,
+  });
 
   final int initialTabIndex;
+  final String? quickFilter; 
 
   @override
   State<InvoiceTabbedPage> createState() => _InvoiceTabbedPageState();
@@ -31,17 +36,19 @@ class _InvoiceTabbedPageState extends State<InvoiceTabbedPage>
     _tabController.dispose();
     super.dispose();
   }
+
   @override
   Widget build(BuildContext context) {
-    final index = _tabController.index;
+    final showBack = widget.quickFilter != null; // filtered route
     return Scaffold(
       appBar: AppBar(
         title: const Text('Invoice Assistant'),
+        leading: showBack ? BackButton(onPressed: () => Navigator.of(context).maybePop()) : null,
         bottom: TabBar(
           unselectedLabelColor: Theme.of(context).secondaryHeaderColor,
           labelColor: Colors.white,
           controller: _tabController,
-          tabs: [
+          tabs: const [
             Tab(icon: Icon(Icons.dashboard_outlined, color: Colors.white), text: 'Dashboard'),
             Tab(icon: Icon(Icons.upload_file_outlined, color: Colors.white), text: 'Upload Invoice'),
             Tab(icon: Icon(Icons.list_alt_outlined, color: Colors.white), text: 'Invoice List'),
@@ -51,9 +58,12 @@ class _InvoiceTabbedPageState extends State<InvoiceTabbedPage>
       body: TabBarView(
         controller: _tabController,
         children: [
-          _DashboardTab(key: const ValueKey('dashboard')),
-          _UploadInvoiceTab(key: const ValueKey('upload')),
-          _InvoiceListTab(key: const ValueKey('list')),
+          const _DashboardTab(key: ValueKey('dashboard')),
+          const _UploadInvoiceTab(key: ValueKey('upload')),
+          _InvoiceListTab(
+            key: const ValueKey('list'),
+            quickFilter: widget.quickFilter, // <— pass it in
+          ),
         ],
       ),
     );
@@ -63,15 +73,11 @@ class _InvoiceTabbedPageState extends State<InvoiceTabbedPage>
 /// Wrapper that renders InvoiceDashboardPage without its Scaffold/AppBar/Drawer
 class _DashboardTab extends StatelessWidget {
   const _DashboardTab({super.key});
-
   @override
   Widget build(BuildContext context) {
-    // We'll use a nested Navigator to intercept the Scaffold
     return Navigator(
       onGenerateRoute: (settings) {
-        return MaterialPageRoute(
-          builder: (context) => const InvoiceDashboardPage(),
-        );
+        return MaterialPageRoute(builder: (context) => const InvoiceDashboardPage());
       },
     );
   }
@@ -80,14 +86,11 @@ class _DashboardTab extends StatelessWidget {
 /// Wrapper that renders UploadInvoicePage without its Scaffold/AppBar/Drawer
 class _UploadInvoiceTab extends StatelessWidget {
   const _UploadInvoiceTab({super.key});
-
   @override
   Widget build(BuildContext context) {
     return Navigator(
       onGenerateRoute: (settings) {
-        return MaterialPageRoute(
-          builder: (context) => const UploadInvoicePage(),
-        );
+        return MaterialPageRoute(builder: (context) => const UploadInvoicePage());
       },
     );
   }
@@ -95,14 +98,15 @@ class _UploadInvoiceTab extends StatelessWidget {
 
 /// Wrapper that renders InvoiceListPage without its Scaffold/AppBar/Drawer
 class _InvoiceListTab extends StatelessWidget {
-  const _InvoiceListTab({super.key});
+  const _InvoiceListTab({super.key, this.quickFilter});
+  final String? quickFilter;
 
   @override
   Widget build(BuildContext context) {
     return Navigator(
       onGenerateRoute: (settings) {
         return MaterialPageRoute(
-          builder: (context) => const InvoiceListPage(),
+          builder: (context) => InvoiceListPage(quickFilter: quickFilter),
         );
       },
     );
